+++ next.config.js
@@ -1,10 +1,10 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
-  output: 'export',
+  output: "export",
   images: {
     unoptimized: true,
   },
   trailingSlash: true,
-}
+};
 
-module.exports = nextConfig
+module.exports = nextConfig;